hack/requirements.txt
docopt
pyyaml
requests